requirements.txt
uvicorn~=0.17.6
fastapi==0.75.0
passlib==1.7.4
pydantic==1.9.0
python_jose==3.3.0
setuptools==59.5.0
tortoise-orm==0.18.1
aiofiles~=0.8.0
python-multipart==0.0.5
requests~=2.27.1